refactor: first god-module split — Mission.FindingOps + API.FindingController
Starts the P2 god-module decomposition (Mission 7014 / ApiController
2328) with the extraction pattern: move a coherent resource surface,
delegate from the god module so every call site keeps working.

- Mission.FindingOps: approve/reject/escalate/dispose_finding/
  dispose_session_findings/dispose_findings/finding_audit_events + the
  atomic status+audit Multi and disposition-filter privates. Mission
  defdelegates to it; the shared emitters (emit_finding_event,
  record_finding_memory, emit_platform_finding_event) became public
  @doc-false functions FindingOps calls back into.
  Mission: 7014 -> 6774 lines.
- ControlKeelWeb.APIHelpers: the api/v1 authorization surface
  (session/finding/task/workspace access, workspace-for-conn scope
  check, integer parsing, changeset errors, finding_summary serializer)
  extracted from ApiController privates; controllers import it so
  existing call sites compile unchanged.
- ControlKeelWeb.API.FindingController: list_findings/create_finding/
  finding_action moved out of ApiController (2328 -> 2140); router
  points the three /findings routes at it.

Same pattern extends to ReviewOps/ProofOps and remaining controllers
as follow-up slices.

mix precommit: 2242/0 (1 excluded), CI guard passed.
